# api/cmd/api/docs/swagger.yaml
definitions:
  github_com_Caqil_megapdf-api_internal_models.Account:
    properties:
      accessToken:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expiresAt:
        type: integer
      id:
        type: integer
      idToken:
        type: string
      provider:
        type: string
      providerAccountId:
        type: string
      refreshToken:
        type: string
      scope:
        type: string
      sessionState:
        type: string
      tokenType:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.ApiKey:
    properties:
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      key:
        type: string
      lastUsed:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
...
  github_com_Caqil_megapdf-api_internal_models.PaymentWebhookEvent:
    properties:
      createdAt:
        type: string
      eventId:
        type: string
      eventType:
        type: string
      id:
        type: string
      processedAt:
        type: string
      rawData:
        type: string
      resourceId:
        type: string
      resourceType:
        type: string
      status:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.Session:
    properties:
      expires:
        type: string
      id:
        type: string
      sessionToken:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.SwaggerUser:
    description: User account information
    properties:
      balance:
        example: 100.5
        type: number
      createdAt:
        type: string
      email:
        example: user@example.com
        type: string
      emailVerified:
        type: string
      freeOperationsReset:
        type: string
      freeOperationsUsed:
        example: 5
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image:
        type: string
      isEmailVerified:
        example: false
        type: boolean
      name:
        example: John Doe
        type: string
      role:
        example: user
        type: string
      updatedAt:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.Transaction:
    properties:
      amount:
        type: number
      balanceAfter:
        type: number
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      paymentId:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.UsageStats:
    properties:
      count:
        type: integer
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      operation:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.User:
    properties:
      accounts:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
        type: array
      apiKeys:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
        type: array
      balance:
        type: number
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: string
      freeOperationsReset:
        type: string
      freeOperationsUsed:
        type: integer
      id:
        type: string
      image:
        type: string
      isEmailVerified:
        type: boolean
      name:
        type: string
      password:
        type: string
      role:
        type: string
      sessions:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
        type: array
      transactions:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
        type: array
      updatedAt:
        type: string
      usageStats:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
        type: array
      verificationToken:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.VerificationToken:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expires:
        type: string
      id:
        type: integer
      identifier:
        type: string
      token:
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.SwaggerUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.SwaggerUser'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieves a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates a user with the provided details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /verification-tokens:
    post:
      consumes:
      - application/json
      description: Creates a new verification token with the provided details
      parameters:
      - description: VerificationToken data
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.VerificationToken'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.VerificationToken'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new verification token
      tags:
      - verification-tokens
  /verification-tokens/{identifier}:
    delete:
      consumes:
      - application/json
      description: Deletes a verification token by its identifier
      parameters:
      - description: VerificationToken Identifier
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a verification token
      tags:
      - verification-tokens
    get:
      consumes:
      - application/json
      description: Retrieves a verification token by its identifier
      parameters:
      - description: VerificationToken Identifier
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.VerificationToken'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a verification token
      tags:
      - verification-tokens
  /webhooks:
    post:
      consumes:
      - application/json
      description: Creates a new webhook event with the provided details
      parameters:
      - description: WebhookEvent data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PaymentWebhookEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PaymentWebhookEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new webhook event
      tags:
      - webhooks
  /webhooks/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a webhook event by its ID
      parameters:
      - description: WebhookEvent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a webhook event
      tags:
      - webhooks
    get:
      consumes:
      - application/json
      description: Retrieves a webhook event by its ID
      parameters:
      - description: WebhookEvent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PaymentWebhookEvent'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a webhook event
      tags:
      - webhooks
    put:
      consumes:
      - application/json
      description: Updates a webhook event with the provided details
      parameters:
      - description: WebhookEvent ID
        in: path
        name: id
        required: true
        type: string
      - description: WebhookEvent data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PaymentWebhookEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PaymentWebhookEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a webhook event
      tags:
      - webhooks
swagger: "2.0"